* generated classes for complete interface
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@171 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/src/org/xtreemfs/interfaces/StripingPolicy.java
@@ -17,19 +17,19 @@ import java.util.ArrayList;
 public class StripingPolicy implements org.xtreemfs.interfaces.utils.Serializable
 {
     public StripingPolicy() { policy = 0; stripe_size = 0; width = 0; }
-    public StripingPolicy( int policy, long stripe_size, long width ) { this.policy = policy; this.stripe_size = stripe_size; this.width = width; }
+    public StripingPolicy( int policy, int stripe_size, int width ) { this.policy = policy; this.stripe_size = stripe_size; this.width = width; }
 
     public int getPolicy() { return policy; }
     public void setPolicy( int policy ) { this.policy = policy; }
-    public long getStripe_size() { return stripe_size; }
-    public void setStripe_size( long stripe_size ) { this.stripe_size = stripe_size; }
-    public long getWidth() { return width; }
-    public void setWidth( long width ) { this.width = width; }
+    public int getStripe_size() { return stripe_size; }
+    public void setStripe_size( int stripe_size ) { this.stripe_size = stripe_size; }
+    public int getWidth() { return width; }
+    public void setWidth( int width ) { this.width = width; }
 
     // Object
     public String toString()
     {
-        return "StripingPolicy( " + Integer.toString( policy ) + ", " + Long.toString( stripe_size ) + ", " + Long.toString( width ) + " )";
+        return "StripingPolicy( " + Integer.toString( policy ) + ", " + Integer.toString( stripe_size ) + ", " + Integer.toString( width ) + " )";
     }    
 
     // Serializable
@@ -37,29 +37,29 @@ public class StripingPolicy implements org.xtreemfs.interfaces.utils.Serializabl
     
     public void serialize(ONCRPCBufferWriter writer) {
         writer.putInt( policy );
-        writer.putLong( stripe_size );
-        writer.putLong( width );        
+        writer.putInt( stripe_size );
+        writer.putInt( width );        
     }
     
     public void deserialize( ReusableBuffer buf )
     {
         policy = buf.getInt();
-        stripe_size = buf.getLong();
-        width = buf.getLong();    
+        stripe_size = buf.getInt();
+        width = buf.getInt();    
     }
     
     public int calculateSize()
     {
         int my_size = 0;
         my_size += ( Integer.SIZE / 8 );
-        my_size += ( Long.SIZE / 8 );
-        my_size += ( Long.SIZE / 8 );
+        my_size += ( Integer.SIZE / 8 );
+        my_size += ( Integer.SIZE / 8 );
         return my_size;
     }
 
     private int policy;
-    private long stripe_size;
-    private long width;
+    private int stripe_size;
+    private int width;
 
 }
